% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_functions.R
\name{calculate_grm}
\alias{calculate_grm}
\title{calculate_grm}
\usage{
calculate_grm(sample_by_gene_matrix)
}
\arguments{
\item{sample_by_gene_matrix}{must have first column as sample names and other columns as gene names}
}
\value{
grm genetic relatedness matrix for tau and beta test
}
\description{
This function alculates genetic relatedness matrix (GRM) from sample-by-gene matrix
}
