% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp_genedata}
\alias{exp_genedata}
\title{Example Binary Gene Presence Absence Data}
\format{
A binary sample-by-gene data frame: with 1 as present, 0 as absent
\describe{
\item{sample_name}{Character. The sample identifier.}
\item{gene1}{Character. Gene ID 1.}
\item{gene2}{Character. Gene ID 2.}
\item{...}{Additional gene IDs.}
}
}
\source{
Generated  from microSLAM
}
\usage{
exp_genedata
}
\description{
This dataset contains input data used in microSLAM analysis.
}
\keyword{datasets}
