% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp_metadata}
\alias{exp_metadata}
\title{Example Metadata}
\format{
A data frame with X rows and Y columns:
\describe{
\item{sample_name}{Character. The sample identifier.}
\item{y}{Numeric. Disease condition (e.g., "healthy", "disease").}
\item{age}{Numeric. Age of the sample}
\item{strain}{Numeric. Disease condition (e.g., "healthy", "disease").}
\item{uncorrelated_strain}{Numeric. Disease condition (e.g., "healthy", "disease").}
}
}
\source{
Generated internally from microSLAM
}
\usage{
exp_metadata
}
\description{
This dataset contains example metadata about samples.
}
\keyword{datasets}
