% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_functions.R
\name{fit_beta}
\alias{fit_beta}
\title{fit_beta}
\usage{
fit_beta(pop.struct.glmm, glm.fit0, grm, sample_by_gene_df, SPA = FALSE)
}
\arguments{
\item{pop.struct.glmm}{output of fit_tau_test; GLMM of species with grm accounted for}

\item{glm.fit0}{glm model. Model output with no sample relatedness accounted for}

\item{grm}{Genetic Relatedness Matrix (from calculate_grm or user) NxN matrix of sample relatedness}

\item{sample_by_gene_df}{sample-by-gene data frame with sample_name, and list_of_genes}

\item{SPA}{whether to run Saddle point approximation for pvalues (will slow down output)}
}
\value{
dataframe of beta estimates for all genes tested
}
\description{
Fit a beta for each genes including the population structure model with the random effects
}
