% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_functions.R
\name{fit_tau_test}
\alias{fit_tau_test}
\title{fit_tau_test}
\usage{
fit_tau_test(
  glm_fit0,
  GRM,
  species_id,
  tau = c(1, 1),
  maxiter = 100,
  verbose = TRUE,
  tol = 1e-04,
  log_file = NA
)
}
\arguments{
\item{glm_fit0}{glm model. Model output with no sample relatedness accounted for}

\item{GRM}{Genetic Relatedness Matrix (from scripts or user) NxN matrix of sample relatedness}

\item{species_id}{for tracking species}

\item{tau}{vector for initial values for the variance component parameter estimates usually c(1,1)}

\item{maxiter}{maximum iterations to fit the glmm model}

\item{verbose}{whether outputting messages in the process of model fitting}

\item{log_file}{log file to write to}
}
\value{
model output for the baseline structure glmm
}
\description{
Fit the base model for population structure finding random effects
}
