% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microSLAM-package.R
\docType{_PACKAGE}
\name{microSLAM-package}
\alias{microSLAM-package}
\alias{microSLAM}
\title{The microSLAM package}
\description{
microSLAM is a library and R package for microbiome GLMM models
}
\keyword{glmm}
