% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_genes.R
\name{prep_genes_function_R}
\alias{prep_genes_function_R}
\title{prep_genes_function_R}
\usage{
prep_genes_function_R(
  gcopynumber,
  gdepth,
  depth_cutoff,
  samples_per_copynumber,
  verbose = FALSE,
  make_plots = FALSE,
  write_csv = FALSE,
  output_dir = NULL,
  s_id = "s_id",
  pangenome_used = FALSE,
  centroid_prevalence_file = NULL,
  centroid_prevalence_cutoff = 0.7,
  genes_info = NULL,
  GRM_used = FALSE,
  GRM = NULL,
  genes_summary_used = FALSE,
  genes_summary = NULL,
  metadata_used = FALSE,
  metadata = NULL,
  min_control_case_ratio = 0,
  log_scale = FALSE,
  mean_center = FALSE,
  is_var_filter = FALSE,
  var_filter = 0
)
}
\arguments{
\item{gcopynumber}{gcopynumber}

\item{gdepth}{gdepth}

\item{depth_cutoff}{depth_cutoff}

\item{samples_per_copynumber}{samples_per_copynumber}
}
\description{
helper function to filter snps data from midas2 and make GRM
}
