% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_snps.R
\name{prep_snps_function_R}
\alias{prep_snps_function_R}
\title{prep_snps_function_R}
\usage{
prep_snps_function_R(
  snp_freq,
  snp_depth,
  snp_info,
  sample_median_depth_filter,
  number_of_samples_for_sites,
  verbose = FALSE,
  make_plots = FALSE,
  pangenome_used = FALSE,
  centroid_prevalence_cutoff = 0.8,
  run_qp = FALSE,
  l = 0.3,
  u = 3,
  a = 5,
  start_samples,
  start_snps,
  genes_summary_used,
  genes_summary
)
}
\arguments{
\item{snp_freq}{snp_freq from MIDAS2}

\item{snp_depth}{snp_depth from MIDAS2}

\item{snp_info}{snp_info from MIDAS2}

\item{sample_median_depth_filter}{sample median depth filter}

\item{number_of_samples_for_sites}{number of samples for sites}
}
\description{
helper function to filter snps data from midas2 and make GRM
}
