% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_functions.R
\name{run_beta_test}
\alias{run_beta_test}
\title{Run Beta test for all genes}
\usage{
run_beta_test(
  sample_by_gene_df,
  glm_fit0,
  glmm_fit,
  GRM,
  genes_to_test,
  do_spa = TRUE
)
}
\arguments{
\item{sample_by_gene_df}{A data frame with gene expression values per sample.}

\item{glm_fit0}{The null GLM model.}

\item{glmm_fit}{The fitted GLMM model.}

\item{GRM}{The genomic relationship matrix (GRM).}

\item{genes_to_test}{A data frame of genes to test.}

\item{do_spa}{Logical. If TRUE, use Saddlepoint Approximation (SPA).}
}
\value{
A data frame with beta test results.
}
\description{
Runs a beta test on all genes using GLM and GLMM models.
}
