% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_functions.R
\name{run_microslam}
\alias{run_microslam}
\title{Run microSLAM Analysis}
\usage{
run_microslam(
  sample_by_gene_to_test,
  samples_to_test,
  genes_to_test,
  GRM_to_test,
  species_id,
  outdir,
  logdir,
  do_SPA = FALSE,
  formula_string = "y ~ age + 1",
  response_var = "y",
  family = "binomial",
  verbose = FALSE,
  n_tau = 1000,
  debug = FALSE
)
}
\arguments{
\item{sample_by_gene_to_test}{A data frame with binary gene presence/absence.}

\item{samples_to_test}{A metadata data frame for samples, with sample_name as first column.}

\item{genes_to_test}{A metadata data frame for genes, with gene_id as first column.}

\item{GRM_to_test}{A data frame for the genomic relationship matrix (GRM).}

\item{species_id}{Character. Species identifier.}

\item{outdir}{Character. Directory for saving output files.}

\item{logdir}{Character. Directory for storing log files.}

\item{do_SPA}{Logical. Whether to run saddlepoint approximation (SPA) for p-value estimation (default: \code{FALSE}).}

\item{formula_string}{Character. Model formula (default = \code{"y ~ age + 1"}).}

\item{response_var}{Character. Name of the response variable (default = \code{"y"}).}

\item{family}{Family in GLM (default = 'binomial')}

\item{verbose}{Logical (default = FALSE)}

\item{n_tau}{Integer. Number of tau permutations (default = \code{1000}).}

\item{debug}{Logical (default = FALSE) Only for developing debug purpose}
}
\value{
A fitted GLMM model (\code{glmm_fit}).
}
\description{
This function runs the microSLAM analysis pipeline, which includes:
\itemize{
\item Tau test for strain-level trait association, with permutation test for p value
\item Beta test for gene-level trait association.
}
}
