% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_functions.R
\name{run_tau_test}
\alias{run_tau_test}
\title{run_tau_test}
\usage{
run_tau_test(glm_fit0, GRM, n_tau, species_id = "s_id", tau0, phi0)
}
\arguments{
\item{glm_fit0}{glm model. Model output with no sample relatedness accounted for}

\item{GRM}{Genetic Relatedness Matrix (from scripts or user) NxN matrix of sample relatedness}

\item{n_tau}{number of tau to simulate}

\item{species_id}{species id for bactrial species}

\item{tau0}{starting tau}

\item{phi0}{starting phi}
}
\value{
df of values of T for tau for different runs
}
\description{
take output from population structure test and test how exterme the tau is for that set of data
}
