% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_functions.R
\name{summary.pop.struct.glmm}
\alias{summary.pop.struct.glmm}
\title{summary.pop.struct.glmm}
\usage{
\method{summary}{pop.struct.glmm}(x, ...)
}
\arguments{
\item{x}{a pop.struct.glmm objecct the output of fit_tau_test; GLMM of species with grm accounted for}

\item{...}{Additional arguments passed to other summary methods.}
}
\description{
Summarize the output of fit tau test
}
