% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_functions.R
\name{summary}
\alias{summary}
\title{... all the usual documentation for summary() ...}
\usage{
summary(x, ...)
}
\arguments{
\item{x}{An object to summarize.}

\item{...}{Additional arguments passed to other summary methods.}
}
\description{
... all the usual documentation for summary() ...
}
