% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{validate_gene_test}
\alias{validate_gene_test}
\title{validate_gene_test}
\usage{
validate_gene_test(opt)
}
\arguments{
\item{opt}{commandline arguments from gene test}
}
\description{
helper function to validate gene test from command line
}
