% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{validate_GRM_metadata}
\alias{validate_GRM_metadata}
\title{validate_GRM_metadata}
\usage{
validate_GRM_metadata(opt)
}
\arguments{
\item{opt}{input from command line}
}
\description{
helper function to validate GRM and metadata
}
