% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_snps.R
\name{validate_snps_input}
\alias{validate_snps_input}
\title{validate_snps_input}
\usage{
validate_snps_input(opt)
}
\arguments{
\item{opt}{the inputs from the command line}
}
\description{
helper function to validate the input from command line for snps
}
