% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_functions.R
\name{calculate_GRM}
\alias{calculate_GRM}
\title{calculate_GRM}
\usage{
calculate_GRM(gene_matrix)
}
\arguments{
\item{genematrix}{must have first column as sample names and other columns as gene names}
}
\value{
GRM
}
\description{
calculate the GRM from the gene matrix
}
