% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microSLAM-package.R
\docType{package}
\name{microSLAM-package}
\alias{microSLAM-package}
\alias{microSLAM}
\title{The microSLAM package}
\description{
microSLAM is a library and R package for performing metagenome-wide association studies with
generalized linear mixed effects models (GLMMs)
}
\author{
\strong{Maintainer}: Miriam Goldman \email{miriamagoldman@gmail.com}

}
\keyword{glmm}
