% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microSLAM-package.R
\docType{package}
\name{microSLAM-package}
\alias{microSLAM-package}
\alias{microSLAM}
\title{The microSLAM package}
\description{
microSLAM is a library and R package for microbiome GLMM models
}
\author{
\strong{Maintainer}: Miriam Goldman \email{miriam.goldman@ucsf.edu}

}
\keyword{glmm}
