% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm_functions.R
\name{run_microslam}
\alias{run_microslam}
\title{Run microSLAM Analysis}
\usage{
run_microslam(
  sample_by_gene_to_test,
  samples_to_test,
  genes_to_test,
  GRM_to_test,
  species_id,
  outdir,
  logdir,
  do_SPA = FALSE,
  formula_string = "y ~ age + 1",
  response_var = "y",
  glm_family = "binomial",
  verbose = TRUE,
  n_tau = 1000
)
}
\arguments{
\item{sample_by_gene_to_test}{A data frame with gene expression values per sample.}

\item{samples_to_test}{A data frame containing sample metadata.}

\item{genes_to_test}{A data frame listing the genes to be tested.}

\item{GRM_to_test}{The genomic relationship matrix (GRM).}

\item{species_id}{Character. Species identifier.}

\item{outdir}{Character. Directory for saving output files.}

\item{logdir}{Character. Directory for storing log files.}

\item{do_SPA}{Run saddle point estimation (default = FALSE)}

\item{formula_string}{Character. Model formula (default = \code{"y ~ age + 1"}).}

\item{response_var}{Character. Name of the response variable (default = \code{"y"}).}

\item{glm_family}{Family in GLM (default = 'binomial')}

\item{verbose}{Logical (default = TRUE)}

\item{n_tau}{Integer. Number of tau permutations (default = \code{1000}).}
}
\value{
A fitted GLMM model (\code{glmm_fit}).
}
\description{
This function runs the microSLAM analysis pipeline, which includes:
\itemize{
\item Tau test for strain-level association
\item Beta test for gene-level association
\item Saving all results to specified directories.
}
}
